requirements.txt
Flask==3.0.3
faiss-cpu==1.10.0
numpy==2.1.3
sentence-transformers==3.4.1
openai==1.37.1
python-dotenv==1.0.1
requests==2.32.3